drivers: virt: gunyah: Add support for input device configuration
Add support for the configuration of an input device. Input devices are
unique and require banking data in the device configuration space. To do
so, two hypercalls are added:
1. input_configure
2. input_set_data

The input_configure call is used to specify the input device's ID,
property bits, and the number of supported events and axes.

The input_set_data call is used to specify the configuration data for
each available event and axes as well as the device's name and unique
serial ID.

Change-Id: I3ba7b01ee7ce462a52b510aa9d4a378e1f4476ba
Signed-off-by: Anant Goel <quic_anantg@quicinc.com>