gunyah: Handle VM restart support when qcrosvm exits
When qcrosvm is killed a signal is pending on the thread.
wait_event_interruptible would exit without waiting for the condition
to be set.Move this to wait_event instead as we need to wait for VMs to
cleanly exit.

Signed-off-by: Anant Goel <quic_anantg@quicinc.com>
Signed-off-by: Sumedha Sanjay Shenolikar <quic_shenolik@quicinc.com>
(cherry picked from commit 1e49e9f30ca065ae5522e0806becb38000a5e37b)

Change-Id: Ib481d251df888196e3c095661f6af97b8213b299