drivers: virt: gunyah: Clean up EventFD and IRQFD context for BE devices
Ensure that a cleanup of the EventFDs and IRQFDs is done on MMIO exit
for all backend devices. This ensures that no contexts or FDs are left
in an open state which prevent subsequent VM restarts.

Signed-off-by: Anant Goel <quic_anantg@quicinc.com>
Signed-off-by: Sumedha Sanjay Shenolikar <quic_shenolik@quicinc.com>
(cherry picked from commit bfa3edf5bb7256336ab776f4b7b349354b10b4d2)

Change-Id: Iea53b45637848d551d79a715fd0ec65f2121beac