virt: gunyah: Add support for querying information for VM Memory Regions
Add support to the Kernel Gunyah interface for VMM to query the VM
memory count and fetch information on the individual VM memory regions.
This information is used by VMM to build Guest memory FDs out of the
memory regions which are used by devices running in a VMM, vHost, or
vHost User context to access the virtqueues and data buffers created
by the driver.

The information comprises of the start address of the region, its size,
and an FD which represents the region.

Change-Id: I89e81b09448fc90d0aee544faab89a551369640a
Signed-off-by: Anant Goel <quic_anantg@quicinc.com>