drivers: virt: gunyah: Handle guest suspend notification
Enhance gunyah.ko to receive the guest suspend notification
from RM and deliver the notification to subscribed user
space clients through uEventD framework.

The Suspend event will be reported to ueventD as "suspended"
and the suspend type would be reported as "vm_suspend_type".
The suspend type would be 2 for VM_STATE_CPU_SUSPENDED
and it would be 3 for VM_STATE_SYSTEM_SUSPENDED.

Change-Id: I1db0b98721cabaca24a67d8893cb3812b64055cb
Signed-off-by: Mahesh Kumar <quic_mahkumar@quicinc.com>