gunyah: Remove PTE_SPECIAL for guest memory
With enabling the O_DIRECT flag on the underlying disk image,
we need to pin user pages, the virtio buffer from guest memory,
and add them to a bio, which doesn't support the memory with
PTE_SPECIAL page table helper. Here we remove PTE_SPECIAL after
mapping the guest memory.

Change-Id: Ib35be138bcb8d4f673b18cbce2b0691c8746e7ea