gunyah-drivers: Add support for IOMEM sharing across VMs
If qcom,iomem flag is set for a shared memory, the memory is
considered as an IO memory region and is allowed to be shared
across VMs which are defined as dst-VMs in the device tree.

Bookkeeping is done to ensure that the MEMSHARE API is invoked
only once per shared IOMEM. Also reclaim the memory if the IOMEM
is not freed by the VMs. Additionally, the reclaiming of IO memory
ensures proper functioning of VMs during restart scenarios.

Change-Id: I8d0ddf6c6f98eca0376f3f1849b044c7b1d653d2
Signed-off-by: Mahesh Kumar <quic_mahkumar@quicinc.com>
Signed-off-by: Xiaoguang Zheng <quic_xiazhe@quicinc.com>
Signed-off-by: Yao Jiang <quic_c_yaojia@quicinc.com>